requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart
pinecone-client
langchain